Make AsyncClient noncopyable/nonmovable
The previous code tried to implement move-as-copy semantics, which can
lead to unexpected behavior; and further it was not safe, as there was a
risk of race conditions with the LwIP thread.

Remove the unsafe function for now, and indicate to the compiler to
check for it.

If there's a strong need for a move operation in the future, the thread
safety can be reviewed at that time.